switched to using int for price and volume

+++ initial/src/main/java/Matcher/HelloController.java
@@ -22,13 +22,13 @@ public class HelloController {
 	}
 
 	@PostMapping("/user")
-	public Account postAccount(@RequestParam String username, double startingGBP, double startingBTC){
+	public Account postAccount(@RequestParam String username, int startingGBP, int startingBTC){
 		matcher.createAccount(username, startingGBP, startingBTC);
 		return matcher.getAccount(username);
 	}
 
 	@PostMapping("/order")
-	public Order postOrder (@RequestParam String username, String action, double volume, double price) {
+	public Order postOrder (@RequestParam String username, String action, int volume, int price) {
 		Order order = new Order(username, action, volume, price);
 		matcher.processOrder(order);
 		return order;
